% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycboroughs.R
\docType{data}
\name{nycboroughs}
\alias{nycboroughs}
\title{NYC Borough Boundaries}
\format{SpatialPolygonsDataFrame with columns
\describe{
\item{shape_leng}{Length of borough}
\item{boroname}{Name of borough}
\item{borocode}{Numerical code of borough}
\item{shape_area}{Area of borough}
}}
\source{
NYC Department of City Planning via NYC OpenData
\url{https://data.cityofnewyork.us/City-Government/Borough-Boundaries/tqmj-j8zm}
}
\usage{
nycboroughs
}
\description{
Spatial data for the boundaries of the five boroughs of New York City: the
Bronx, Brooklyn, Manhattan, Queens, and Staten Island.
}
\keyword{datasets}

